terrain: move biter_texture to precomputed/ directory
As there are more and more machine generated files, move those to
dedicated directory.
