+++ maps/biter_battles_v2/precomputed/biter_texture.lua
